GH-1: don't send BSpace upon Vim misidentification
This reverts commit 7d45311e478bcb86f26dec38aff8c1befd88d2c0 to fix #1.

It seems that an older version of the navigation logic (lost to history)
needed this code.  The current logic only looks at the beginning of the
pane title, and is therefore not confused by "Vim" appearing elsewhere.

For example, Vim 7.2 sets the pane title to "Thanks for flying Vim" and
doesn't bother to clean up after itself upon termination.  As a result,
the old navigation logic would still see "Thanks for flying Vim" as the
title and think that Vim was still running: therefore misidentification.